Feature/read configs

+++ program/file_manager.py
@@ -1,4 +1,5 @@
 import json
+import os
 import shutil
 
 class file_manager:
@@ -23,22 +24,43 @@ class file_manager:
         self.config_file = open(self._CONFIG_FILE_, "a+")
         self.config_file.seek(0)
 
-        if not self.read_config(config_name):
-            self.write_config(config_name)
+        if os.stat(self._CONFIG_FILE_).st_size > 0:
+            configs = json.load(self.config_file)
+            selected_config = configs.get(config_name)
+
+            if not selected_config:
+                self.write_config(config_name, configs)
+                return
+
+            self.source_files = selected_config["source_files"]
+            self.source_directories = selected_config["source_directories"]
+            self.output_directories = selected_config["output_directories"]
+        else:
+            self.write_config(config_name, "")
 
         self.config_file.close()
+        
+    def write_config(self, new_config_name, existing_configs):
+        if not self.source_directories and not self.source_files:
+            print("ERROR: please specify at least one source file (-f) or source directory (-d)")
+            return
 
-    def read_config(self, config_name):
-        for line in self.config_file:
-            print(line)
-        return False
-    
-    def write_config(self, config_name):
+        if not self.output_directories:
+            print("ERROR: please specify at least one output directory (-o)")
+            return
+
         new_config = {
-            "name": config_name,
-            "source_files": self.source_files,
-            "source_directories": self.source_directories,
-            "output_directories": self.output_directories
-        }
-        json.dump(new_config, self.config_file, indent=4)
-        return
+                        new_config_name: {
+                        "source_files": self.source_files,
+                        "source_directories": self.source_directories,
+                        "output_directories": self.output_directories
+                    }
+                }
+
+        if existing_configs == "":
+            json.dump(new_config, self.config_file, indent=4)
+            return
+        existing_configs.update(new_config)
+        self.config_file.close()
+        self.config_file = open(self._CONFIG_FILE_, "w")
+        json.dump(existing_configs, self.config_file, indent=4)
